Webserver

+++ RTWWebServer/Cache/IRemoteCacheKeyGenerator.cs
@@ -3,5 +3,5 @@ namespace RTWWebServer.Cache;
 public interface IRemoteCacheKeyGenerator
 {
     string GenerateUserLockKey(int userId);
-    string GenerateUserSessionKey(int userId);
+    string GenerateUserSessionKey(long userId);
 }
